task palindrome int, update variables types.

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -7,7 +7,7 @@
  */
 int is_palindrome(unsigned long n)
 {
-	int x, y;
+	unsigned long x, y;
 	unsigned long div = 1;
 
 	if (n < 10)
